Enhance Pattern Matching and Type System
- Updated the grammar to support tuple types in type annotations, improving type expressiveness.
- Enhanced the analysis system with comprehensive pattern exhaustiveness checking, including detailed error reporting for missing patterns.
- Improved the parser to handle tuple type parsing and added validation for minimum tuple element requirements.
- Refactored the analysis logic to provide better feedback on non-exhaustive patterns and added tests to ensure robustness.

This change significantly enhances the type system and pattern matching capabilities, allowing for more precise type definitions and improved error handling during analysis.

+++ src/internal/tests.rs
@@ -461,13 +461,13 @@ mod tests {
 
         match result {
             Value::List(items) => {
-                // Note: Since ThreadSafeFunction is a placeholder, the actual filtering/mapping
-                // behavior may not be fully implemented yet
-                assert_eq!(items.len(), 4); // All items pass through with placeholder
-                assert_eq!(items[0], Value::Integer(1));
-                assert_eq!(items[1], Value::Integer(2));
-                assert_eq!(items[2], Value::Integer(3));
-                assert_eq!(items[3], Value::Integer(4));
+                // Map operation doubles values: [1,2,3,4] -> [2,4,6,8]
+                // Filter operation keeps even numbers: [2,4,6,8] -> [2,4,6,8] (all are even)
+                assert_eq!(items.len(), 4); // All mapped values pass filter
+                assert_eq!(items[0], Value::Integer(2)); // 1 * 2 = 2 (even, passes filter)
+                assert_eq!(items[1], Value::Integer(4)); // 2 * 2 = 4 (even, passes filter)
+                assert_eq!(items[2], Value::Integer(6)); // 3 * 2 = 6 (even, passes filter)
+                assert_eq!(items[3], Value::Integer(8)); // 4 * 2 = 8 (even, passes filter)
             }
             _ => assert!(false, "Expected LazyValue::MapFiltered, got: {:?}", result),
         }
